v13 pre add stace trace

+++ app/src/main/java/book/course/molareza/ir/mp3player/activity/FileDownloader.java
@@ -6,6 +6,7 @@ import android.os.AsyncTask;
 import android.util.Log;
 import android.widget.Toast;
 
+import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
@@ -15,19 +16,19 @@ import java.net.MalformedURLException;
 import java.net.URL;
 
 import book.course.molareza.ir.mp3player.G;
+import book.course.molareza.ir.mp3player.MyToast;
 
-/**
- * Created by Home on 6/21/2016.
- */
+
 public class FileDownloader extends AsyncTask<String, Integer, String> {
 
     public static ProgressDialog mProgres;
 
     private Context context;
 
+    private boolean isCheck ;
     public FileDownloader(Context cont) {
 
-        this.context =cont;
+        this.context = cont;
 
         mProgres = new ProgressDialog(context);
         mProgres.setMessage("فایل در حال دانلود می باشد");
@@ -63,63 +64,73 @@ public class FileDownloader extends AsyncTask<String, Integer, String> {
         String[] nameUrl = params[0].split("/");
         String nameFile = nameUrl[nameUrl.length - 1];
 
-        try {
-            URL url = new URL(params[0]);
-            connection = (HttpURLConnection) url.openConnection();
-            connection.setDoInput(true);
-            connection.connect();
+        File check = new File(params[1] + nameFile);
 
-            int fileSize = connection.getContentLength();
+        if (!check.exists()) {
 
-            inputStream = connection.getInputStream();
-
-            outputStream = new FileOutputStream(params[1] + nameFile);
-
-            byte[] buffer = new byte[4096];
-
-            long percent = 0;
-            int len;
-
-            while ((len = inputStream.read(buffer)) != -1) {
-
-                percent += len;
-
-                if (fileSize > 0) {
-
-                    publishProgress((int) (percent * 100 / fileSize));
-
-                    Log.i("TAGTAGTAG", "doInBackground: " + percent);
-
-                }
-                outputStream.write(buffer, 0, len);
-            }
-
-
-        } catch (MalformedURLException e) {
-            e.printStackTrace();
-        } catch (IOException e) {
-            e.printStackTrace();
-        } finally {
+            isCheck = true;
 
             try {
-                if (outputStream !=null){
-                    outputStream.flush();
-                    outputStream.close();
+                URL url = new URL(params[0]);
+                connection = (HttpURLConnection) url.openConnection();
+                connection.setDoInput(true);
+                connection.connect();
 
+                int fileSize = connection.getContentLength();
+
+                inputStream = connection.getInputStream();
+
+                outputStream = new FileOutputStream(params[1] + nameFile);
+
+                byte[] buffer = new byte[4096];
+
+                long percent = 0;
+                int len;
+
+                while ((len = inputStream.read(buffer)) != -1) {
+
+                    percent += len;
+
+                    if (fileSize > 0) {
+
+                        publishProgress((int) (percent * 100 / fileSize));
+
+                        Log.i("TAGTAGTAG", "doInBackground: " + percent);
+
+                    }
+                    outputStream.write(buffer, 0, len);
                 }
-                if (inputStream !=null){
 
-                    inputStream.close();
-                }
 
+            } catch (MalformedURLException e) {
+                e.printStackTrace();
             } catch (IOException e) {
                 e.printStackTrace();
+            } finally {
+
+                try {
+                    if (outputStream != null) {
+                        outputStream.flush();
+                        outputStream.close();
+
+                    }
+                    if (inputStream != null) {
+
+                        inputStream.close();
+                    }
+
+                } catch (IOException e) {
+                    e.printStackTrace();
+                }
+
+                if (connection != null) {
+
+                    connection.disconnect();
+                }
             }
 
-            if (connection !=null){
-
-                connection.disconnect();
-            }
+        } else {
+           isCheck = false;
         }
 
         return null;
@@ -129,15 +140,20 @@ public class FileDownloader extends AsyncTask<String, Integer, String> {
     protected void onPostExecute(String s) {
         super.onPostExecute(s);
 
-        mProgres.dismiss();
+        if (isCheck){
+            mProgres.dismiss();
 
-        Log.i("TAGTAGTAG", "t7676t76t: "  + G.DIR_CLIP);
-        if (s != null) {
+            if (s != null) {
 
-            Toast.makeText(G.context, "خطا در حین دانلود", Toast.LENGTH_SHORT).show();
-        } else {
-            Toast.makeText(G.context, "دانلود با موفقیت صورت گرفت", Toast.LENGTH_SHORT).show();
+                MyToast.makeText(G.context, "خطا در حین دانلود", Toast.LENGTH_SHORT).show();
+            } else {
+                MyToast.makeText(G.context, "دانلود با موفقیت صورت گرفت", Toast.LENGTH_SHORT).show();
 
+            }
+        }else {
+            mProgres.dismiss();
+            MyToast.makeText(G.context, "این فایل قبلا دانلود شده", Toast.LENGTH_SHORT).show();
         }
+
     }
 }
